Allow creating draft if wrong method is selected
This is supported because in some systems an order can be created from
a backend where the module does not have a control over which method
is available, so an admin can add any method. Therefore, it can happen
that an admin adds a method that does not support
delivery to the given destination country.

Issue: PACKLINK-25